+++ NEWS.md
@@ -1,6 +1,14 @@
 Core Hunter 3
 =============
 
+Version 3.1.0 (27/01/2017)
+--------------------------
+
+ - More informative output when printing Core Hunter data objects (#19, #23).
+ - It is now required to explicitly specify the format when reading genotype data. Moreover, a warning is raised in case it seems that the wrong format may have been selected, through an inspection of the data that was read (#20).
+ - Fixed issues when reading files with single or double quoted values.
+ - Improved test coverage.
+
 Version 3.0.1 (04/10/2016)
 --------------------------